fix(main): pass real BigHeader* to LOAD_LangFile

+++ game/MAIN/MainMain.c
@@ -596,7 +596,7 @@ void StateZero()
 
 	// English=1
 	// PAL SCES02105 calls it multiple times
-	LOAD_LangFile((int)sdata->ptrBigfile1, 1);
+	LOAD_LangFile(sdata->ptrBigfile1, 1);
 	GAMEPROG_NewGame_OnBoot();
 	gGT->overlayIndex_null_notUsed = 0;
